v2.7 update
add feature import parallel
fix bug: 6 digit microseconds value is not handled for mysql time data type
