+++ source code/TestDebugASPNET/masterPage1.Master
@@ -31,6 +31,7 @@
             <a href="/ShaFilesCompute">SHA256 Files Checksum Compute</a>
             <a href="/CreateSampleTableRows">Create Sample Table Rows</a>
             <a href="/ParallelPipelineExport">Parallel Pipeline Export</a>
+            <a href="/MySqlDateTimeTester">MySqlDateTime Tester</a>
             <a href="/Maintenance">Maintenance</a>
             <a href="/Typography">Typography</a>
             <a href="/About">About</a>
