+++ source code/TestDebugASPNET/pages/ReportProgress.aspx.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.UI;
+using System.Web.UI.WebControls;
+
+namespace System.pages
+{
+    public partial class ReportProgress : System.Web.UI.Page
+    {
+        protected void Page_Load(object sender, EventArgs e)
+        {
+            Response.Redirect("~/ProgressReport1", true);
+        }
+    }
+}
